backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["willidragon <willidragon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
openai = "^1.40.3"
langchain = "^0.2.12"
langchain-community = "^0.2.11"
langchain-openai = "^0.1.21"
python-decouple = "^3.8"
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


